+++ src/util.hpp
@@ -61,6 +61,7 @@ namespace globed_util {
     CCScene* sceneWithLayer(CCNode* layer);
     void handleErrors();
     bool isNumeric(const std::string& str);
+    void loadDeathEffects();
 
     inline void errorPopup(const std::string& text) {
         FLAlertLayer::create("Globed Error", text, "Ok")->show();